gray summon boat when not near water
